Refactored code out of ValidationManagerUI.
