+++ Validation-Manager-Web/src/main/java/net/sourceforge/javydreamercsw/validation/manager/web/ValidationManagerUI.java
@@ -24,8 +24,6 @@ import com.vaadin.annotations.VaadinServletConfiguration;
 import com.vaadin.data.Item;
 import com.vaadin.data.Property;
 import com.vaadin.data.fieldgroup.BeanFieldGroup;
-import com.vaadin.data.fieldgroup.FieldGroupFieldFactory;
-import com.vaadin.data.util.BeanItemContainer;
 import com.vaadin.data.util.HierarchicalContainer;
 import com.vaadin.event.FieldEvents.TextChangeEvent;
 import com.vaadin.event.ItemClickEvent;
@@ -38,13 +36,10 @@ import com.vaadin.server.Page;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.server.VaadinServlet;
 import com.vaadin.shared.MouseEventDetails.MouseButton;
-import com.vaadin.shared.ui.datefield.Resolution;
 import com.vaadin.shared.ui.dd.VerticalDropLocation;
 import com.vaadin.ui.AbstractSelect;
-import com.vaadin.ui.Alignment;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.CheckBox;
-import com.vaadin.ui.ComboBox;
 import com.vaadin.ui.Component;
 import com.vaadin.ui.Field;
 import com.vaadin.ui.FormLayout;
@@ -57,11 +52,9 @@ import com.vaadin.ui.Label;
 import com.vaadin.ui.Layout;
 import com.vaadin.ui.Notification;
 import com.vaadin.ui.Panel;
-import com.vaadin.ui.PopupDateField;
 import com.vaadin.ui.TabSheet;
 import com.vaadin.ui.TabSheet.Tab;
 import com.vaadin.ui.TextArea;
-import com.vaadin.ui.TextField;
 import com.vaadin.ui.Tree;
 import com.vaadin.ui.Tree.TreeDragMode;
 import com.vaadin.ui.Tree.TreeDropCriterion;
@@ -89,33 +82,21 @@ import com.validation.manager.core.db.Project;
 import com.validation.manager.core.db.Requirement;
 import com.validation.manager.core.db.RequirementSpec;
 import com.validation.manager.core.db.RequirementSpecNode;
-import com.validation.manager.core.db.RequirementSpecPK;
-import com.validation.manager.core.db.SpecLevel;
 import com.validation.manager.core.db.Step;
 import com.validation.manager.core.db.TestCase;
 import com.validation.manager.core.db.TestCaseExecution;
 import com.validation.manager.core.db.TestPlan;
 import com.validation.manager.core.db.TestProject;
 import com.validation.manager.core.db.VmSetting;
-import com.validation.manager.core.db.VmUser;
 import com.validation.manager.core.db.controller.ExecutionStepJpaController;
 import com.validation.manager.core.db.controller.ProjectJpaController;
-import com.validation.manager.core.db.controller.RequirementJpaController;
-import com.validation.manager.core.db.controller.RequirementSpecJpaController;
-import com.validation.manager.core.db.controller.RequirementSpecNodeJpaController;
-import com.validation.manager.core.db.controller.SpecLevelJpaController;
 import com.validation.manager.core.db.controller.StepJpaController;
 import com.validation.manager.core.db.controller.TestCaseExecutionJpaController;
-import com.validation.manager.core.db.controller.TestCaseJpaController;
-import com.validation.manager.core.db.controller.TestPlanJpaController;
-import com.validation.manager.core.db.controller.TestProjectJpaController;
 import com.validation.manager.core.db.controller.exceptions.IllegalOrphanException;
 import com.validation.manager.core.db.controller.exceptions.NonexistentEntityException;
 import com.validation.manager.core.history.Versionable;
 import com.validation.manager.core.history.Versionable.CHANGE_LEVEL;
-import com.validation.manager.core.server.core.ExecutionStepServer;
 import com.validation.manager.core.server.core.ProjectServer;
-import com.validation.manager.core.server.core.RequirementServer;
 import com.validation.manager.core.server.core.TestCaseExecutionServer;
 import com.validation.manager.core.server.core.TestCaseServer;
 import com.validation.manager.core.server.core.VMSettingServer;
@@ -135,17 +116,14 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Properties;
@@ -158,13 +136,21 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.annotation.WebServlet;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.BaselineComponent;
-import net.sourceforge.javydreamercsw.validation.manager.web.component.ByteToStringConverter;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.ExecutionStepComponent;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.HistoryTable;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.ProjectTreeComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.RequirementComponent;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.RequirementListComponent;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.RequirementSelectionComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.RequirementSpecComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.RequirementSpecNodeComponent;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.StepComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.TestCaseComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.TestCaseExecutionComponent;
 import net.sourceforge.javydreamercsw.validation.manager.web.component.TestCaseExporter;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.TestPlanComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.TestProjectComponent;
+import net.sourceforge.javydreamercsw.validation.manager.web.component.VMWindow;
 import net.sourceforge.javydreamercsw.validation.manager.web.dashboard.ExecutionDashboard;
 import net.sourceforge.javydreamercsw.validation.manager.web.demo.DemoProvider;
 import net.sourceforge.javydreamercsw.validation.manager.web.importer.FileUploader;
@@ -277,97 +263,8 @@ public class ValidationManagerUI extends UI implements VMUI {
 
     private void displayRequirementSpecNode(RequirementSpecNode rsn,
             boolean edit) {
-        Panel form = new Panel();
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(rsn.getClass());
-        binder.setItemDataSource(rsn);
-        Field<?> name = binder.buildAndBind(TRANSLATOR.translate("general.name"),
-                "name");
-        layout.addComponent(name);
-        Field desc = binder.buildAndBind(TRANSLATOR.translate("general.description"),
-                "description",
-                TextArea.class);
-        desc.setSizeFull();
-        layout.addComponent(desc);
-        Field<?> scope = binder.buildAndBind(TRANSLATOR.translate("general.scope"),
-                "scope");
-        layout.addComponent(scope);
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (rsn.getRequirementSpecNodePK() == null) {
-                displayObject(rsn.getRequirementSpec());
-            } else {
-                displayObject(rsn, false);
-            }
-        });
-        if (edit) {
-            if (rsn.getRequirementSpecNodePK() == null) {
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        rsn.setName(name.getValue().toString());
-                        rsn.setDescription(desc.getValue().toString());
-                        rsn.setScope(scope.getValue().toString());
-                        rsn.setRequirementSpec((RequirementSpec) tree.getValue());
-                        new RequirementSpecNodeJpaController(DataBaseManager
-                                .getEntityManagerFactory()).create(rsn);
-                        form.setVisible(false);
-                        //Recreate the tree to show the addition
-                        updateProjectList();
-                        buildProjectTree(rsn);
-                        displayObject(rsn, true);
-                        updateScreen();
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    rsn.setName(name.getValue().toString());
-                    rsn.setDescription(desc.getValue().toString());
-                    rsn.setScope(scope.getValue().toString());
-                    handleVersioning(rsn, () -> {
-                        try {
-                            new RequirementSpecNodeJpaController(DataBaseManager
-                                    .getEntityManagerFactory()).edit(rsn);
-                            displayRequirementSpecNode(rsn, true);
-                        } catch (NonexistentEntityException ex) {
-                            LOG.log(Level.SEVERE, null, ex);
-                            Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                    ex.getLocalizedMessage(),
-                                    Notification.Type.ERROR_MESSAGE);
-                        } catch (Exception ex) {
-                            LOG.log(Level.SEVERE, null, ex);
-                            Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                    ex.getLocalizedMessage(),
-                                    Notification.Type.ERROR_MESSAGE);
-                        }
-                    });
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        binder.setReadOnly(!edit);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, REQUIREMENT_REVIEW);
+        setTabContent(main, new RequirementSpecNodeComponent(rsn, edit),
+                REQUIREMENT_REVIEW);
     }
 
     public void setTabContent(Tab target, Component content,
@@ -395,280 +292,12 @@ public class ValidationManagerUI extends UI implements VMUI {
 
     private void displayTestCaseExecution(TestCaseExecution tce,
             ProjectServer ps, boolean edit) {
-        Panel form = new Panel(TRANSLATOR.translate("test.case.execution.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(tce.getClass());
-        binder.setItemDataSource(tce);
-        Field<?> name = binder.buildAndBind(TRANSLATOR.translate("general.name"),
-                "name");
-        name.setRequired(true);
-        name.setRequiredError(TRANSLATOR.translate("missing.name.message"));
-        layout.addComponent(name);
-        Field<?> scope = binder.buildAndBind(TRANSLATOR.translate("general.scope"),
-                "scope");
-        layout.addComponent(scope);
-        //TODO: Show when finished
-        TextArea conclusion = new TextArea(TRANSLATOR.translate("general.conclusion"));
-        binder.bind(conclusion, "conclusion");
-        layout.addComponent(conclusion);
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (tce.getId() == null) {
-                displayObject(tree.getValue());
-            } else {
-                displayObject(tce, false);
-            }
-        });
-        if (edit) {
-            if (tce.getId() == null) {
-                TestCaseExecutionServer tces = new TestCaseExecutionServer();
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    if (name.getValue() == null) {
-                        Notification.show(name.getRequiredError(),
-                                Notification.Type.ERROR_MESSAGE);
-                        return;
-                    }
-                    Map<Requirement, History> history = new HashMap<>();
-                    if (ps != null) {
-                        List<Requirement> toApprove = new ArrayList<>();
-                        Tool.extractRequirements(ps).forEach((r) -> {
-                            //Check each requirement and see if they have minor versions (last version is not baselined)
-                            History h = r.getHistoryList().get(r.getHistoryList().size() - 1);
-                            if (h.getMajorVersion() == 0
-                                    || h.getMidVersion() > 0
-                                    || h.getMinorVersion() > 0) {
-                                if (r.getHistoryList().size() == 1) {
-                                    //Nothing to choose from
-                                    history.put(r, h);
-                                } else {
-                                    toApprove.add(r);
-                                }
-                            } else {
-                                history.put(r, h);
-                            }
-                        });
-                        if (!toApprove.isEmpty()) {
-                            MessageBox mb = MessageBox.create();
-                            mb.asModal(true)
-                                    .withCaption(TRANSLATOR.translate("missing.baseline.requirement.title"))
-                                    .withMessage(TRANSLATOR.translate("missing.baseline.requirement.message"))
-                                    .withButtonAlignment(Alignment.MIDDLE_CENTER)
-                                    .withOkButton(() -> {
-                                        //Start the wizard
-                                        Wizard w = new Wizard();
-                                        VMWindow sw = new VMWindow();
-                                        w.setDisplayedMaxTitles(3);
-                                        toApprove.forEach(r -> {
-                                            w.addStep(new SelectRequirementVersionStep(r));
-                                        });
-                                        w.addListener(new WizardProgressListener() {
-                                            @Override
-                                            public void activeStepChanged(WizardStepActivationEvent event) {
-                                                //Do nothing
-                                            }
-
-                                            @Override
-                                            public void stepSetChanged(WizardStepSetChangedEvent event) {
-                                                //Do nothing
-                                            }
-
-                                            @Override
-                                            public void wizardCompleted(WizardCompletedEvent event) {
-                                                //Process the selections
-                                                w.getSteps().forEach(s -> {
-                                                    SelectRequirementVersionStep step
-                                                            = (SelectRequirementVersionStep) s;
-                                                    history.put(step.getRequirement(),
-                                                            step.getHistory());
-                                                });
-                                                removeWindow(sw);
-                                            }
-
-                                            @Override
-                                            public void wizardCancelled(WizardCancelledEvent event) {
-                                                removeWindow(sw);
-                                            }
-                                        });
-                                        sw.setContent(w);
-                                        sw.setSizeFull();
-                                        addWindow(sw);
-                                    }, ButtonOption.focus(),
-                                            ButtonOption.icon(VaadinIcons.CHECK))
-                                    .withCancelButton(
-                                            ButtonOption.icon(VaadinIcons.CLOSE)
-                                    ).getWindow().setIcon(ValidationManagerUI.SMALL_APP_ICON);
-                            mb.open();
-                        }
-                    }
-                    if (!history.isEmpty()) {
-                        try {
-                            if (conclusion.getValue() != null) {
-                                tces.setConclusion(conclusion.getValue());
-                            }
-                            if (scope.getValue() != null) {
-                                tces.setScope(scope.getValue().toString());
-                            }
-                            tces.setName(name.getValue().toString());
-                            tces.write2DB();
-                            if (ps != null) {
-                                //Process the list
-                                ps.getTestProjects(true).forEach(tp -> {
-                                    tces.addTestProject(tp);
-                                });
-                                //Now look thru the ExecutionSteps and assign the right version.
-                                tces.getExecutionStepList().forEach(es -> {
-                                    try {
-                                        ExecutionStepServer ess = new ExecutionStepServer(es);
-                                        es.getStep().getRequirementList().forEach(r -> {
-                                            ess.getHistoryList().add(history.get(r));
-                                        });
-                                        ess.write2DB();
-                                    } catch (VMException ex) {
-                                        LOG.log(Level.SEVERE, null, ex);
-                                    }
-                                });
-                            }
-                            tces.write2DB();
-                            tces.update(tce, tces.getEntity());
-                            updateProjectList();
-                            updateScreen();
-                            displayObject(tce);
-                        } catch (Exception ex) {
-                            LOG.log(Level.SEVERE, null, ex);
-                        }
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                TestCaseExecutionServer tces = new TestCaseExecutionServer(tce);
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    tces.setConclusion(conclusion.getValue());
-                    tces.setScope(scope.getValue().toString());
-                    tces.setName(name.getValue().toString());
-                    try {
-                        handleVersioning(tces, () -> {
-                            try {
-                                tces.write2DB();
-                                tces.update(tce, tces.getEntity());
-                                displayObject(tce);
-                            } catch (NonexistentEntityException ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            } catch (Exception ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            }
-                        });
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        binder.setReadOnly(!edit);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, "testcase.view");
+        setTabContent(main, new TestCaseExecutionComponent(tce, ps,
+                edit), "testcase.view");
     }
 
     private void displayExecutionStep(ExecutionStep es) {
-        Panel form = new Panel(TRANSLATOR.translate("execution.step.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(es.getClass());
-        binder.setItemDataSource(es);
-        FieldGroupFieldFactory defaultFactory = binder.getFieldFactory();
-        binder.setFieldFactory(new FieldGroupFieldFactory() {
-
-            @Override
-            public <T extends Field> T createField(Class<?> dataType, Class<T> fieldType) {
-                if (dataType.isAssignableFrom(VmUser.class)) {
-                    BeanItemContainer<VmUser> userEntityContainer
-                            = new BeanItemContainer<>(VmUser.class);
-                    userEntityContainer.addBean(es.getAssignee());
-                    Field field = new TextField(es.getAssignee() == null
-                            ? TRANSLATOR.translate("general.not.applicable")
-                            : es.getAssignee().getFirstName() + " "
-                            + es.getAssignee().getLastName());
-                    return fieldType.cast(field);
-                }
-
-                return defaultFactory.createField(dataType, fieldType);
-            }
-        });
-        layout.addComponent(createStepHistoryTable(TRANSLATOR.translate("step.detail"),
-                Arrays.asList(es.getStepHistory()), false));
-        if (es.getResultId() != null) {
-            Field<?> result = binder.buildAndBind(TRANSLATOR.translate("general.result"),
-                    "resultId.resultName");
-            layout.addComponent(result);
-        }
-        if (es.getComment() != null) {
-            TextArea comment = new TextArea(TRANSLATOR.translate("general.comment"));
-            binder.bind(comment, "comment");
-            layout.addComponent(comment);
-        }
-        if (es.getAssignee() != null) {
-            TextField assignee = new TextField(TRANSLATOR.translate("general.assignee"));
-            VmUser u = es.getAssignee();
-            assignee.setValue(u.toString());
-            assignee.setReadOnly(true);
-            layout.addComponent(assignee);
-        }
-        if (es.getExecutionStart() != null) {
-            Field<?> start = binder.buildAndBind(TRANSLATOR.translate("execution.start"),
-                    "executionStart");
-            layout.addComponent(start);
-        }
-        if (es.getExecutionEnd() != null) {
-            Field<?> end = binder.buildAndBind(TRANSLATOR.translate("execution.end"),
-                    "executionEnd");
-            layout.addComponent(end);
-        }
-        if (es.getExecutionTime() != null && es.getExecutionTime() > 0) {
-            Field<?> time = binder.buildAndBind(TRANSLATOR.translate("execution.time"),
-                    "executionTime");
-            layout.addComponent(time);
-        }
-        if (!es.getHistoryList().isEmpty()) {
-            layout.addComponent(createRequirementHistoryTable(
-                    TRANSLATOR.translate("related.requirements"),
-                    es.getHistoryList()));
-        }
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (es.getExecutionStepPK() == null) {
-                displayObject(tree.getValue());
-            } else {
-                displayObject(es, false);
-            }
-        });
-        binder.setReadOnly(true);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, "testcase.view");
+        setTabContent(main, new ExecutionStepComponent(es), "testcase.view");
     }
 
     public void displayStep(Step s, boolean edit) {
@@ -676,440 +305,22 @@ public class ValidationManagerUI extends UI implements VMUI {
     }
 
     private void displayTestCase(TestCase t, boolean edit) {
-        Panel form = new Panel(TRANSLATOR.translate("test.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(t.getClass());
-        binder.setItemDataSource(t);
-        Field<?> name = binder.buildAndBind(TRANSLATOR.translate("general.name"),
-                "name");
-        layout.addComponent(name);
-        TextArea summary = new TextArea(TRANSLATOR.translate("general.summary"));
-        summary.setConverter(new ByteToStringConverter());
-        binder.bind(summary, "summary");
-        layout.addComponent(summary);
-        PopupDateField creation = new PopupDateField(TRANSLATOR.translate("general.creation.date"));
-        creation.setResolution(Resolution.SECOND);
-        creation.setDateFormat(VMSettingServer.getSetting("date.format")
-                .getStringVal());
-        binder.bind(creation, "creationDate");
-        layout.addComponent(creation);
-        Field<?> active = binder.buildAndBind(TRANSLATOR.translate("general.active"),
-                "active");
-        layout.addComponent(active);
-        Field<?> open = binder.buildAndBind(TRANSLATOR.translate("general.open"),
-                "isOpen");
-        layout.addComponent(open);
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (t.getId() == null) {
-                displayObject(tree.getValue());
-            } else {
-                displayObject(t, false);
-            }
-        });
-        if (edit) {
-            if (t.getId() == null) {
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        t.setName(name.getValue().toString());
-                        t.setSummary(summary.getValue().getBytes("UTF-8"));
-                        t.setCreationDate((Date) creation.getValue());
-                        t.setActive((Boolean) active.getValue());
-                        t.setIsOpen((Boolean) open.getValue());
-                        t.getTestPlanList().add((TestPlan) tree.getValue());
-                        new TestCaseJpaController(DataBaseManager
-                                .getEntityManagerFactory()).create(t);
-                        form.setVisible(false);
-                        //Recreate the tree to show the addition
-                        updateProjectList();
-                        buildProjectTree(t);
-                        displayTestCase(t, false);
-                        updateScreen();
-                    } catch (UnsupportedEncodingException ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        t.setName(name.getValue().toString());
-                        t.setSummary(summary.getValue().getBytes("UTF-8"));
-                        t.setCreationDate((Date) creation.getValue());
-                        t.setActive((Boolean) active.getValue());
-                        t.setIsOpen((Boolean) open.getValue());
-                        handleVersioning(t, () -> {
-                            try {
-                                new TestCaseJpaController(DataBaseManager
-                                        .getEntityManagerFactory()).edit(t);
-                                displayTestCase(t, true);
-                            } catch (NonexistentEntityException ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            } catch (Exception ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            }
-                        });
-                    } catch (UnsupportedEncodingException ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        binder.setReadOnly(!edit);
-        creation.setEnabled(false);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, "testcase.view");
+        setTabContent(main, new TestCaseComponent(t, edit), "testcase.view");
     }
 
     private void displayTestPlan(TestPlan tp, boolean edit) {
-        Panel form = new Panel(TRANSLATOR.translate("test.plan.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(tp.getClass());
-        binder.setItemDataSource(tp);
-        Field<?> name = binder.buildAndBind(TRANSLATOR.translate("general.name"),
-                "name");
-        layout.addComponent(name);
-        Field notes = binder.buildAndBind(TRANSLATOR.translate("general.notes"),
-                "notes",
-                TextArea.class);
-        notes.setSizeFull();
-        layout.addComponent(notes);
-        Field<?> active = binder.buildAndBind(TRANSLATOR.translate("general.active"),
-                "active");
-        layout.addComponent(active);
-        Field<?> open = binder.buildAndBind(TRANSLATOR.translate("general.open"),
-                "isOpen");
-        layout.addComponent(open);
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (tp.getTestPlanPK() == null) {
-                displayObject(tree.getValue());
-            } else {
-                displayObject(tp, false);
-            }
-        });
-        if (edit) {
-            if (tp.getTestPlanPK() == null) {
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        tp.setName(name.getValue().toString());
-                        tp.setNotes(notes.getValue().toString());
-                        tp.setActive((Boolean) active.getValue());
-                        tp.setIsOpen((Boolean) open.getValue());
-                        new TestPlanJpaController(DataBaseManager
-                                .getEntityManagerFactory()).create(tp);
-                        form.setVisible(false);
-                        //Recreate the tree to show the addition
-                        updateProjectList();
-                        buildProjectTree(tp);
-                        displayTestPlan(tp, false);
-                        updateScreen();
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        tp.setName(name.getValue().toString());
-                        tp.setNotes(notes.getValue().toString());
-                        tp.setActive((Boolean) open.getValue());
-                        tp.setIsOpen((Boolean) open.getValue());
-                        handleVersioning(tp, () -> {
-                            try {
-                                new TestPlanJpaController(DataBaseManager
-                                        .getEntityManagerFactory()).edit(tp);
-                                displayTestPlan(tp, true);
-                            } catch (NonexistentEntityException ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            } catch (Exception ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            }
-                        });
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        binder.setReadOnly(!edit);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, "testcase.view");
+        setTabContent(main, new TestPlanComponent(tp, edit), "testcase.view");
     }
 
     private void displayTestProject(TestProject tp, boolean edit) {
-        Panel form = new Panel(TRANSLATOR.translate("test.project.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(tp.getClass());
-        binder.setItemDataSource(tp);
-        Field<?> name = binder.buildAndBind(TRANSLATOR.translate("general.name"),
-                "name");
-        layout.addComponent(name);
-        Field notes = binder.buildAndBind(TRANSLATOR.translate("general.notes"),
-                "notes",
-                TextArea.class);
-        notes.setSizeFull();
-        layout.addComponent(notes);
-        Field<?> active = binder.buildAndBind(TRANSLATOR.translate("general.active"),
-                "active");
-        layout.addComponent(active);
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (tp.getId() == null) {
-                displayObject(tree.getValue());
-            } else {
-                displayObject(tp, false);
-            }
-        });
-        if (edit) {
-            if (tp.getId() == null) {
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        tp.setName(name.getValue().toString());
-                        tp.setNotes(notes.getValue().toString());
-                        tp.setActive((Boolean) active.getValue());
-                        new TestProjectJpaController(DataBaseManager
-                                .getEntityManagerFactory()).create(tp);
-                        form.setVisible(false);
-                        //Recreate the tree to show the addition
-                        updateProjectList();
-                        buildProjectTree(tp);
-                        displayTestProject(tp, false);
-                        updateScreen();
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        tp.setName(name.getValue().toString());
-                        tp.setNotes(notes.getValue().toString());
-                        tp.setActive((Boolean) active.getValue());
-                        handleVersioning(tp, () -> {
-                            try {
-                                new TestProjectJpaController(DataBaseManager
-                                        .getEntityManagerFactory()).edit(tp);
-                                displayTestProject(tp, true);
-                            } catch (NonexistentEntityException ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            } catch (Exception ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            }
-                        });
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        binder.setReadOnly(!edit);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, "testcase.view");
+        setTabContent(main, new TestProjectComponent(tp, edit), "testcase.view");
     }
 
     private void displayRequirementSpec(RequirementSpec rs, boolean edit) {
-        Panel form = new Panel(TRANSLATOR.translate("requirement.spec.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(rs.getClass());
-        binder.setItemDataSource(rs);
-        Field<?> name = binder.buildAndBind(TRANSLATOR.translate("general.name"), "name");
-        layout.addComponent(name);
-        Field desc = binder.buildAndBind(TRANSLATOR.translate("general.description"), "description",
-                TextArea.class);
-        desc.setSizeFull();
-        layout.addComponent(desc);
-        Field<?> date = binder.buildAndBind(TRANSLATOR.translate("general.modification.data"),
-                "modificationDate");
-        layout.addComponent(date);
-        date.setEnabled(false);
-        SpecLevelJpaController controller
-                = new SpecLevelJpaController(DataBaseManager
-                        .getEntityManagerFactory());
-        List<SpecLevel> levels = controller.findSpecLevelEntities();
-        BeanItemContainer<SpecLevel> specLevelContainer
-                = new BeanItemContainer<>(SpecLevel.class, levels);
-        ComboBox level = new ComboBox(TRANSLATOR.translate("spec.level"));
-        level.setContainerDataSource(specLevelContainer);
-        level.getItemIds().forEach(id -> {
-            level.setItemCaption(id, TRANSLATOR.translate(((SpecLevel) id).getName()));
-        });
-        binder.bind(level, "specLevel");
-        layout.addComponent(level);
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (rs.getRequirementSpecPK() == null) {
-                displayObject(rs.getProject());
-            } else {
-                displayObject(rs, false);
-            }
-        });
-        if (edit) {
-            if (rs.getRequirementSpecPK() == null) {
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        rs.setName(name.getValue().toString());
-                        rs.setModificationDate(new Date());
-                        rs.setSpecLevel((SpecLevel) level.getValue());
-                        rs.setProject(((Project) tree.getValue()));
-                        rs.setRequirementSpecPK(new RequirementSpecPK(
-                                rs.getProject().getId(),
-                                rs.getSpecLevel().getId()));
-                        new RequirementSpecJpaController(DataBaseManager
-                                .getEntityManagerFactory()).create(rs);
-                        form.setVisible(false);
-                        //Recreate the tree to show the addition
-                        updateProjectList();
-                        buildProjectTree(rs);
-                        displayObject(rs, true);
-                        updateScreen();
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.creation"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        rs.setName(name.getValue().toString());
-                        rs.setModificationDate(new Date());
-                        rs.setSpecLevel((SpecLevel) level.getValue());
-                        handleVersioning(rs, () -> {
-                            try {
-                                new RequirementSpecJpaController(DataBaseManager
-                                        .getEntityManagerFactory()).edit(rs);
-                                displayRequirementSpec(rs, true);
-                            } catch (NonexistentEntityException ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            } catch (Exception ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            }
-                        });
-                    } catch (Exception ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        binder.setReadOnly(!edit);
-        binder.bindMemberFields(form);
-        layout.setSizeFull();
-        form.setSizeFull();
-        setTabContent(main, form, REQUIREMENT_REVIEW);
+        setTabContent(main, new RequirementSpecComponent(rs, edit), REQUIREMENT_REVIEW);
     }
 
+    @Override
     public void displayObject(Object item) {
         displayObject(item, false);
     }
@@ -1168,137 +379,7 @@ public class ValidationManagerUI extends UI implements VMUI {
     }
 
     private void displayRequirement(Requirement req, boolean edit) {
-        Panel form = new Panel(TRANSLATOR.translate("requirement.detail"));
-        FormLayout layout = new FormLayout();
-        form.setContent(layout);
-        form.addStyleName(ValoTheme.FORMLAYOUT_LIGHT);
-        BeanFieldGroup binder = new BeanFieldGroup(req.getClass());
-        binder.setItemDataSource(req);
-        Field<?> id = binder.buildAndBind(TRANSLATOR.translate("requirement.id"),
-                "uniqueId");
-        layout.addComponent(id);
-        Field desc = binder.buildAndBind(TRANSLATOR.translate("general.description"),
-                "description",
-                TextArea.class);
-        desc.setSizeFull();
-        layout.addComponent(desc);
-        Field notes = binder.buildAndBind(TRANSLATOR.translate("general.notes"),
-                "notes",
-                TextArea.class);
-        notes.setSizeFull();
-        layout.addComponent(notes);
-        if (req.getParentRequirementId() != null) {
-            TextField tf = new TextField(TRANSLATOR.translate("general.parent"));
-            tf.setValue(req.getParentRequirementId().getUniqueId());
-            tf.setReadOnly(true);
-            layout.addComponent(tf);
-        }
-        if (!req.getRequirementList().isEmpty() && !edit) {
-            layout.addComponent(getDisplayRequirementList(
-                    TRANSLATOR.translate("related.requirements"),
-                    req.getRequirementList()));
-        } else if (edit) {
-            //Allow user to add children
-            AbstractSelect as = getRequirementSelectionComponent();
-            req.getRequirementList().forEach(sub -> {
-                as.select(sub);
-            });
-            as.addValueChangeListener(event -> {
-                Set<Requirement> selected
-                        = (Set<Requirement>) event.getProperty().getValue();
-                req.getRequirementList().clear();
-                selected.forEach(r -> {
-                    req.getRequirementList().add(r);
-                });
-            });
-            layout.addComponent(as);
-        }
-        Button cancel = new Button(TRANSLATOR.translate("general.cancel"));
-        cancel.addClickListener((Button.ClickEvent event) -> {
-            binder.discard();
-            if (req.getId() == null) {
-                displayObject(req.getRequirementSpecNode());
-            } else {
-                displayRequirement(req, false);
-            }
-        });
-        if (edit) {
-            if (req.getId() == null) {
-                //Creating a new one
-                Button save = new Button(TRANSLATOR.translate("general.save"));
-                save.addClickListener((Button.ClickEvent event) -> {
-                    req.setUniqueId(id.getValue().toString());
-                    req.setNotes(notes.getValue().toString());
-                    req.setDescription(desc.getValue().toString());
-                    req.setRequirementSpecNode((RequirementSpecNode) tree
-                            .getValue());
-                    new RequirementJpaController(DataBaseManager
-                            .getEntityManagerFactory()).create(req);
-                    form.setVisible(false);
-                    //Recreate the tree to show the addition
-                    buildProjectTree(req);
-                    displayObject(req, true);
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(save);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            } else {
-                //Editing existing one
-                Button update = new Button(TRANSLATOR.translate("general.update"));
-                update.addClickListener((Button.ClickEvent event) -> {
-                    try {
-                        RequirementServer rs = new RequirementServer(req);
-                        rs.setDescription(((TextArea) desc).getValue());
-                        rs.setNotes(((TextArea) notes).getValue());
-                        rs.setUniqueId(((TextField) id).getValue());
-                        handleVersioning(rs, () -> {
-                            try {
-                                rs.write2DB();
-                                //Recreate the tree to show the addition
-                                buildProjectTree(rs.getEntity());
-                                displayRequirement(rs.getEntity(), false);
-                            } catch (NonexistentEntityException ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            } catch (Exception ex) {
-                                LOG.log(Level.SEVERE, null, ex);
-                                Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                        ex.getLocalizedMessage(),
-                                        Notification.Type.ERROR_MESSAGE);
-                            }
-                        });
-                        form.setVisible(false);
-                    } catch (VMException ex) {
-                        LOG.log(Level.SEVERE, null, ex);
-                        Notification.show(TRANSLATOR.translate("general.error.record.update"),
-                                ex.getLocalizedMessage(),
-                                Notification.Type.ERROR_MESSAGE);
-                    }
-                });
-                HorizontalLayout hl = new HorizontalLayout();
-                hl.addComponent(update);
-                hl.addComponent(cancel);
-                layout.addComponent(hl);
-            }
-        }
-        try {
-            //Add a history section
-            List<History> versions = new RequirementServer(req).getHistoryList();
-            if (!versions.isEmpty()) {
-                layout.addComponent(createRequirementHistoryTable(
-                        TRANSLATOR.translate("general.history"), versions));
-            }
-        } catch (VMException ex) {
-            LOG.log(Level.SEVERE, null, ex);
-        }
-        binder.setBuffered(true);
-        binder.setReadOnly(!edit);
-        binder.bindMemberFields(form);
-        form.setSizeFull();
-        setTabContent(main, form, REQUIREMENT_REVIEW);
+        setTabContent(main, new RequirementComponent(req, edit), REQUIREMENT_REVIEW);
     }
 
     // @return the current application instance
@@ -1679,7 +760,7 @@ public class ValidationManagerUI extends UI implements VMUI {
                                             LOG.log(Level.SEVERE, null, ex);
                                         }
                                     }
-                                    buildProjectTree(new TestCaseServer(tc.getId())
+                                    buildProjectTree(new TestCaseServer(tc.getTestCasePK())
                                             .getEntity());
                                     updateScreen();
                                 } catch (TestCaseImportException ex) {
@@ -2413,7 +1494,8 @@ public class ValidationManagerUI extends UI implements VMUI {
                 REQUIREMENT_REVIEW);
     }
 
-    private Component createStepHistoryTable(String title,
+    @Override
+    public Component createStepHistoryTable(String title,
             List<History> historyItems, boolean showVersionFields) {
         Grid grid = new HistoryTable(title, historyItems, null,
                 showVersionFields,
@@ -2427,9 +1509,11 @@ public class ValidationManagerUI extends UI implements VMUI {
         return grid;
     }
 
+    @Override
     public Component createRequirementHistoryTable(String title,
-            List<History> historyItems) {
-        Grid grid = new HistoryTable(title, historyItems, "uniqueId", true,
+            List<History> historyItems, boolean showVersionFields) {
+        Grid grid = new HistoryTable(title, historyItems, "uniqueId",
+                showVersionFields,
                 "uniqueId", "description", "notes");
         Grid.Column uniqueId = grid.getColumn("uniqueId");
         uniqueId.setHeaderCaption(TRANSLATOR.translate("unique.id"));
@@ -2457,6 +1541,7 @@ public class ValidationManagerUI extends UI implements VMUI {
         return new RequirementSelectionComponent(getParentProject());
     }
 
+    @Override
     public Component getDisplayRequirementList(String title,
             List<Requirement> requirementList) {
         return new RequirementListComponent(title, requirementList);
@@ -2482,7 +1567,7 @@ public class ValidationManagerUI extends UI implements VMUI {
                                 boolean canDelete = true;
                                 for (ExecutionStep es : tces.getExecutionStepList()) {
                                     if (tc == null || Objects.equals(es.getStep().getTestCase()
-                                            .getId(), tc.getId())) {
+                                            .getTestCasePK(), tc.getTestCasePK())) {
                                         if (es.getResultId() != null
                                         && es.getResultId().getResultName()
                                                 .equals("result.pending")) {
@@ -2681,6 +1766,7 @@ public class ValidationManagerUI extends UI implements VMUI {
         }
     }
 
+    @Override
     public void handleVersioning(Object o, Runnable r) {
         if (o instanceof Versionable) {
             Versionable ao = (Versionable) o;
